hw_13/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
redis>=4.2.0
python-dotenv
aioredis
sqlmodel
asyncpg
